requirements.txt
openai>=1.66.0
anthropic>=0.40.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pytest>=8.0.0
pytest-mock>=3.12.0
